--- pages/allFiles.js
@@ -1,15 +1,0 @@
-import MainHeader from "./header/mainHeader";
-import Content from "./main/content";
-import Intro from "./main/intro";
-import Lecturer from "./main/lecturer";
-import Footer from "./footer/footer";
-import Header from './header/header';
-import LecturerInfo from './lecturer/lecturerInfo'
-import Contents from "./contents/contents";
-import Lego from "./contents/lego";
-import Colored_header from "./header/colored_header";
-import Detail from './lecturer/detail';
-import DetailModal from "./lecturer/detailModal";
-
-
-export { MainHeader, Content, Intro, Lecturer, Footer, Header, LecturerInfo, Contents, Lego, Colored_header, Detail, DetailModal}
